feat: 完善用户管理功能和雪花算法集成
- 更新用户模型，使用雪花算法生成用户ID
- 修改用户服务和仓库以支持int64类型的用户ID
- 在主程序中初始化雪花算法节点
- 更新TODO文档，标记雪花ID功能为完成

+++ backend/internal/repository/user_repository.go
@@ -41,14 +41,14 @@ func (r *UserRepository) Update(user *models.User) error {
 }
 
 // Delete 删除用户
-func (r *UserRepository) Delete(id uint) error {
+func (r *UserRepository) Delete(id int64) error {
 	return r.db.Delete(&models.User{}, id).Error
 }
 
 
 
 // FindByID 根据 ID 查找用户
-func (r *UserRepository) FindByID(id uint) (*models.User, error) {
+func (r *UserRepository) FindByID(id int64) (*models.User, error) {
 	var user models.User
 	err := r.db.First(&user, id).Error
 	if err != nil {
